# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Gastos",
        "description": "Api para gestionar gastos de grupos",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "paths": {
        "/categoria/{id}": {
            "get": {
                "operationId": "busquedaCategoria",
                "tags": [
                    "/Categoria"
                ],
                "description": "Obtienes las categorias",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id para obtener una Categoria",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/categoria": {
            "post": {
                "operationId": "creacionCategoria",
                "tags": [
                    "/Categoria"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                  ],
                "produces": [
                    "application/json"
                  ],
                "description": "Crear nueva categoria",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "nombre",
                        "description": "Nombre categoria",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Error al guardar en BBDD"
                    }
                }
            }
        }
    }
}